requirements.txt
cryptography==39.0.1
Pillow==9.2.0
numpy==1.23.3